+++ lib/src/screens/network_selection.dart/widgets/network_selection_screen.dart
@@ -43,7 +43,7 @@ class _NetworkSelectionScreenState extends State<NetworkSelectionScreen> {
         }
         if (widget.viewModel.machineVisibleNetworks.isEmpty) {
           return NoContentWidget(
-              icon: const Icon(Icons.error, color: Colors.red),
+              icon: Icon(Icons.error, color: Theme.of(context).colorScheme.error),
               buttons: [
                 FilledButton(
                   onPressed: () {
@@ -106,13 +106,13 @@ class _NetworkSelectionScreenState extends State<NetworkSelectionScreen> {
                           leading: Icon(
                             widget.viewModel.signalToIcon(widget.viewModel.machineVisibleNetworks[index].signal),
                             size: 24.0,
-                            color: Colors.grey,
+                            color: Theme.of(context).colorScheme.onSurface,
                           ),
                           add: false,
                           trailing: Icon(
                             widget.viewModel.securityToIcon(widget.viewModel.machineVisibleNetworks[index].security),
                             size: 20.0,
-                            color: Colors.grey,
+                            color: Theme.of(context).colorScheme.onSurface,
                           ),
                         ),
                       );
@@ -126,7 +126,6 @@ class _NetworkSelectionScreenState extends State<NetworkSelectionScreen> {
                 child: Center(
                   child: ElevatedButton(
                     style: ElevatedButton.styleFrom(
-                      backgroundColor: Theme.of(context).colorScheme.primary,
                       shape: RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(24.0),
                       ),
@@ -143,7 +142,6 @@ class _NetworkSelectionScreenState extends State<NetworkSelectionScreen> {
                     },
                     child: Text(
                       "My network isn't showing up",
-                      style: TextStyle(color: Theme.of(context).colorScheme.onPrimary, fontWeight: FontWeight.w500),
                     ),
                   ),
                 ),